make waterbutled callback after addons archiving is finished
