+++ tests/utils.py
@@ -20,7 +20,7 @@ from osf_tests.factories import DraftRegistrationFactory
 from osf.models import Sanction, NotificationType, Notification
 from tests.base import get_default_metaschema
 from website.archiver import ARCHIVER_SUCCESS
-from website.archiver import listeners as archiver_listeners
+from website.archiver.tasks import archive_callback
 from website import settings as website_settings
 from osf import features
 
@@ -155,7 +155,7 @@ def mock_archive(project, schema=None, auth=None, draft_registration=None, paren
         # Ensure patches actually apply:
         with mock.patch.object(root_job, 'archive_tree_finished', mock.Mock(return_value=True)), \
              mock.patch('website.archiver.tasks.archive_success.delay', mock.Mock()):
-            archiver_listeners.archive_callback(registration)
+            archive_callback(registration)
 
     if autoapprove:
         sanction = registration.root.sanction
