+++ website/project/signals.py
@@ -9,5 +9,3 @@ write_permissions_revoked = signals.signal('write-permissions-revoked')
 node_deleted = signals.signal('node-deleted')
 
 after_create_registration = signals.signal('post-create-registration')
-
-archive_callback = signals.signal('archive-callback')
